Move inline add row to table extras
