+++ vuu-ui/packages/vuu-data-editing/package.json
@@ -13,9 +13,6 @@
   },
   "dependencies": {
     "@salt-ds/core": "1.67.0",
-    "@salt-ds/styles": "0.2.1",
-    "@salt-ds/window": "0.1.1",
-    "@vuu-ui/vuu-ui-controls": "2.2.1",
     "@vuu-ui/vuu-utils": "2.2.1",
     "@vuu-ui/vuu-utils2": "2.2.1"
   },
