+++ vuu-ui/packages/vuu-table-extras/src/index.ts
@@ -35,6 +35,7 @@ export {
 } from "./datasource-stats/DatasourceStats";
 export { FreezeControl } from "./freeze-control/FreezeControl";
 export { FrozenBanner } from "./freeze-control/FrozenBanner";
+export * from "./inline-add-row";
 export { TabbedTableConfigPanel } from "./tabbed-table-config-panel/TabbedTableConfigPanel";
 export { TabbedTableSettingsAction } from "./tabbed-table-config-panel/TabbedTableSettingsAction";
 export { TableFooter, TableFooterTray } from "./table-footer/TableFooter";
