+++ vuu-ui/packages/vuu-table-extras/src/inline-add-row/index.ts
@@ -1,0 +1,1 @@
+export { InlineAddRow, type InlineAddRowProps } from "./InlineAddRow";
